Update Win64 support

+++ vm/cpu-x86.hpp
@@ -42,7 +42,7 @@ inline static void *get_call_target(cell return_address)
 inline static void set_call_target(cell return_address, void *target)
 {
 	check_call_site(return_address);
-	*(int *)(return_address - 4) = ((cell)target - return_address);
+	*(int *)(return_address - 4) = (u32)((cell)target - return_address);
 }
 
 inline static bool tail_call_site_p(cell return_address)
